fix: vacancies states; phone validation user; submodules
