+++ backend/eventos/serializers/category_events.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from eventos.models.category_events import Category
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = ['id', 'name']  # Agregá otros campos si lo necesitás
