+++ backend/eventos/serializers/shifts.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from eventos.models.shifts import Shift
+
+class ShiftSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Shift
+        fields = ['start_time', 'end_time', 'start_date', 'end_date', 'payment', 'quantity']
