+++ backend/eventos/urls.py
@@ -1,5 +1,8 @@
 from django.urls import path
-from eventos.views.event import CategoryListView, CreateEventView, CreateVacancyView, EventDetailView, EventListView, JobTypeListView, UpcomingEventsView
+from eventos.views.event import CreateEventView, EventDetailView, EventListView, UpcomingEventsView
+from eventos.views.vacancy import CreateVacancyView
+from eventos.views.category_events import CategoryListView
+from eventos.views.job_types import JobTypeListView
 
 urlpatterns = [
 
@@ -12,6 +15,7 @@ urlpatterns = [
 
     # Vacantes 
     path('vacancies/create/', CreateVacancyView.as_view(), name='create-vacancy'),
+    # path('vacancies/list/', VacancyListView.as_view(), name='list-vacancies'),
     path('vacancies/job-types/', JobTypeListView.as_view(), name='list-job-types'),
 
 ]
