+++ backend/user_auth/models/user.py
@@ -37,7 +37,7 @@ class CustomUserManager(BaseUserManager):
 
 class CustomUser(AbstractUser):
     email = models.EmailField(unique=True, null=False, blank=False)
-    phone = models.CharField(max_length=20, null=False, blank=False)
+    phone = models.CharField(max_length=20, null=False, blank=False, unique=True)
 
 
     ROLE_CHOICES = [
