+++ backend/user_auth/serializers/employee.py
@@ -5,6 +5,7 @@ from user_auth.models.user import CustomUser
 from user_auth.models.employee import EmployeeProfile # CAMBIAR ESTO
 from django.contrib.auth.hashers import make_password
 from django.contrib.auth.models import Group
+from django.db.models import Q
 
 class EmployeeRegisterSerializer(serializers.Serializer):
     email = serializers.EmailField()
@@ -23,6 +24,11 @@ class EmployeeRegisterSerializer(serializers.Serializer):
     def validate_dni(self, value):
         if EmployeeProfile.objects.filter(dni=value).exists():
             raise serializers.ValidationError("DNI is already in use.")
+        return value
+    
+    def validate_phone(self, value):
+        if CustomUser.objects.filter(phone=value).exists():
+            raise serializers.ValidationError("Phone is already in use.")
         return value
 
     def create(self, validated_data):
@@ -74,6 +80,9 @@ class CompleteEmployeeSocialSerializer(serializers.Serializer):
 
         if EmployeeProfile.objects.filter(dni=data['dni']).exists():
             raise serializers.ValidationError("DNI is already in use.")
+    
+        if CustomUser.objects.filter(Q(phone=data['phone']) & ~Q(id=user.id)).exists():
+            raise serializers.ValidationError("Phone is already in use.")
 
         return data
